Correct Dates

+++ notificaciones/areas/views.py
@@ -1,3 +1,5 @@
+from pyexpat.errors import messages
+
 from django.shortcuts import render, redirect, get_object_or_404
 from .models import Area
 from .forms import AreaForm
@@ -17,8 +19,18 @@ def lista_areas(request):
     print(f"Rol: {request.user.rol}")
     print(f"Área del usuario: {request.user.area}")
 
+    from notificaciones.notificadores.models import Notificador
+    from notificaciones.bitacora.models import Bitacora
 
-    areas = Area.objects.all()
+
+@rol_requerido(['developer', 'admin', 'coordinador', 'secretaria'])
+def lista_areas(request):    
+    areas = Area.objects.all().order_by('id')
+
+    # Calcular si se puede eliminar. Para cada área, determinar si hay registros relacionados
+    for area in areas:
+        area.puede_eliminar = not area.tiene_notificadores()  # True si NO tiene notificadores    
+
     # Si es secretaria, solo ve su propia área
     if request.user.rol == 'secretaria' and request.user.area:
         areas = areas.filter(id=request.user.area.id)
@@ -52,7 +64,7 @@ def editar_area(request, pk):
         form = AreaForm(request.POST, instance=area)
         if form.is_valid():
             form.save()
-            return redirect('areas:lista')        
+            return redirect('areas:lista_areas')        
     else:
         form = AreaForm(instance=area)
     return render(request, 'areas/formulario_area.html', {'form': form, 'titulo': 'Modificar Área', 'area': area})
@@ -60,7 +72,20 @@ def editar_area(request, pk):
 # 4. ELIMINAR ÁREA
 def eliminar_area(request, pk):
     area = get_object_or_404(Area, pk=pk)
+
+    # Verificar dependencias antes de eliminar
+    if area.tiene_dependencias():
+        messages.error(request, f'No se puede eliminar el área "{area.nombre}" porque tiene registros asociados: {area.get_dependencias_texto()}')
+        return redirect('areas:lista_areas')
+
+
     if request.method == 'POST':
         area.delete()
         return redirect('areas:lista_areas')
+    
+    context = {
+        'area': area,
+    }
+
+
     return render(request, 'areas/confirmar_eliminar.html', {'area': area})
